% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{beta_to_mvalue}
\alias{beta_to_mvalue}
\title{DNA methylation M-values}
\usage{
beta_to_mvalue(data)
}
\arguments{
\item{data}{A numeric vector or matrix of DNAm array beta intensity values.}
}
\value{
A numeric vector or matrix of M-values.
}
\description{
Apply M-value transform to DNA methylation (DNAm) array beta intensity values.
}
