% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{cooks_outlier_pval_histogram}
\alias{cooks_outlier_pval_histogram}
\title{P-value histogram for outlier features}
\usage{
cooks_outlier_pval_histogram(
  dds_fit,
  dds_results,
  f_threshold = 99,
  fill_color = "gray30",
  fill_alpha = 1
)
}
\arguments{
\item{dds_fit}{A DESeqDataSet object after model fitting.}

\item{dds_results}{A DESeqResults object derived from \code{dds_fit}.}

\item{f_threshold}{Percentile of the F distribution used as the cutoff for calling outliers.}

\item{fill_color}{Plot bar color.}

\item{fill_alpha}{Plot bar color alpha value.}
}
\value{
A ggplot object.
}
\description{
Plot p-value histogram for Cook's distance-based feature outliers.
}
\details{
Obtains differential expression test p-values for features flagged as outliers based on a
Cook's distance cutoff derived from the user-specified \code{f_threshold}. The F distribution that
corresponds to \code{f_threshold} is parameterized by the differential exppression regression model
design.
}
