% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{counts.DGEList}
\alias{counts.DGEList}
\title{Get counts from DGEList object}
\usage{
counts.DGEList(object, normalized = F)
}
\arguments{
\item{normalized}{A boolean denoting if counts should be normalized.}

\item{dge_data}{A DGEList object.}
}
\value{
A numeric matrix.
}
\description{
Get counts from DGEList object
}
