% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{filter_dds_features}
\alias{filter_dds_features}
\title{Filter lowly expressed features from a DESeq2DataSet (deprecated)}
\usage{
filter_dds_features(
  dds,
  count_cutoff,
  min_sample_fraction = NULL,
  threshold_variable = NULL,
  use_min_fraction = T,
  approximate = F
)
}
\arguments{
\item{dds}{The DDS object from which to filter features.}

\item{count_cutoff}{The minimum number of counts required for a feature for greater than
\code{min_sample_fraction} proportion of samples to retain the feature.}

\item{min_sample_fraction}{The minimum proportion of samples with greater than \code{count_cutoff}
counts to retain a feature.}

\item{threshold_variable}{Name of the binary variable to use for calculating sample fractions.}

\item{use_min_fraction}{If \code{TRUE}, uses the smaller of the sample fractions calculated from the
\code{threshold_variable}. Otherwise uses the larger of the sample fractions.}

\item{approximate}{If \code{TRUE}, round \code{min_sample_fraction} to a whole number ending in '0' or '5'.}
}
\value{
A DDS object.
}
\description{
Removes expression features below a given threshold from a DESeq2DataSet object.
}
\details{
Included for backwards compatibility only. Use \code{low_expression_filter} instead.
Filters out features considered lowly expressed from a DDS object using a count threshold and
sample proportion set by the user. The sample proportion can also be estimated by calculating
proportions of values from a binary variable specified by the user.
}
