% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_utils.R
\name{filter_table}
\alias{filter_table}
\title{Exclude table dimenions by matching values.}
\usage{
filter_table(data, dimension, by, pattern, trim_whitespace = T)
}
\arguments{
\item{data}{Data frame to subset.}

\item{dimension}{The dimension from which to pull values for matching. '1' for rows and '2' for
columns. The dimension not specified is the one that will be reduced in length.}

\item{by}{The row/column name or index to use to pull values for matching.}

\item{pattern}{The pattern to match against. Can be any fixed string, \code{NA}, or \code{NULL} values.}

\item{trim_whitespace}{If \code{TRUE}, remove leading and trailing whitespace from values to match.}
}
\value{
A data frame with one of the dimensions reduced to exclude rows/columns that match
against the user-specified value.
}
\description{
Subset table to exclude rows/columns that match a specific value.
}
\details{
Uses a user-specified row or column to find matches to a user-specified value. For matches on a
row, corresponding columns are removed. For matches on a column, corresponding rows are removed.
}
