% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{get_cooks_cutoff}
\alias{get_cooks_cutoff}
\title{Get Cook's distance cutoff}
\usage{
get_cooks_cutoff(dds_fit, f_threshold = 99)
}
\arguments{
\item{dds_fit}{A DESeqDataSet object after model fitting.}

\item{f_threshold}{Percentile of the F distribution used as the cutoff for calling outliers.}
}
\value{
A Cook's distance value.
}
\description{
Determine Cook's distance outlier cutoff value.
}
\details{
Uses differential expression regression model design to specify an F distribution and calculate
a value that equates to the user-specified F distribution percentile. This value can be used as
a Cook's distance cutoff above which features would be considered outliers.
}
\seealso{
\code{\link{get_cooks_outliers}}, \code{\link{plot_cooks_per_sample}}
}
