% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{get_cooks_outliers}
\alias{get_cooks_outliers}
\title{Flag outliers using Cook's distance}
\usage{
get_cooks_outliers(dds_fit, f_threshold = 99)
}
\arguments{
\item{dds_fit}{A DESeqDataSet object after model fitting.}

\item{f_threshold}{Percentile of the F distribution used as the cutoff for calling outliers.}
}
\value{
A boolean vector indicating if a feature was influenced by outlier samples.
}
\description{
Flag expression features as outliers based on Cook's distance.
}
\details{
Uses differential expression regression model design to specify an F distribution and calculate
a value that equates to the user-specified F distribution percentile. This value is used as a
Cook's distance cutoff above which features are considered outliers. Since a Cook's distance
value is calculated for every feature and sample, the max Cook's distance value across
samples for each feature is used to call outliers.
}
\seealso{
\code{\link{get_cooks_cutoff}}, \code{\link{plot_cooks_per_sample}}
}
