% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{get_design_mat}
\alias{get_design_mat}
\title{Create design matrix}
\usage{
get_design_mat(
  data,
  var_names = NULL,
  corr_cutoff = 0.75,
  corr_method = c("spearman", "pearson"),
  check_binary_vars = T,
  strings_as_factors = T
)
}
\arguments{
\item{data}{A data frame of potential explanatory variables for the design matrix. Rows should
be records and columns should be variables.}

\item{var_names}{A vector of column names from \code{data} to consider for the design matrix.}

\item{corr_cutoff}{A numeric value in \link{0,1} denoting the absolute correlation cutoff to
classify pairs of collinear variables.}

\item{corr_method}{A string indicating the type of correlation to use for determining
collinearity. Must be \code{pearson} or \code{spearman}.}

\item{check_binary_vars}{A boolean indiciating if dichotomous variables should be included in
the collinearity pruning.}

\item{strings_as_factors}{A boolean indiciating if character variables should be converted to
factors in the returned data frame.}
}
\value{
A data frame of explanatory variables for regression.
}
\description{
Format data table to produce a design matrix.
}
\details{
Creates a design matrix and prunes variables with no variance. Removes pairwise collinear
variables.
}
\seealso{
\code{\link{get_residual_matrix}}
}
