% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode_utils.R
\name{get_gencode_rrna_ids}
\alias{get_gencode_rrna_ids}
\title{Retrieve ribosomal RNA feature IDs}
\usage{
get_gencode_rrna_ids(gtf, feature_type = c("transcript", "gene", "exon"))
}
\arguments{
\item{gtf}{A data frame with data from an imported GENCODE GTF file.}

\item{feature_type}{A string denoting the feature type for which to retrieve records. This
corresponds to the 3rd column values in a GTF file.}
}
\value{
A string vector with GENCODE IDs
}
\description{
Gets ribosomal RNA feature IDs from a GENCODE GTF-format data frame.
}
