% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{get_matrix_r2}
\alias{get_matrix_r2}
\title{Calculate (adjusted) r-squared for a matrix}
\usage{
get_matrix_r2(outcomes, residuals, adjusted = F, dof = NULL)
}
\arguments{
\item{outcomes}{A data frame like object with outcome variables as rows and observations as
columns.}

\item{residuals}{A numeric matrix with each column representing the residuals for an outcome.
The ordering follows the row order of \code{outcomes}.}

\item{adjusted}{A boolean indicating if the adjusted r-squared should be calculated in addition
to the unadjusted r-squared.}

\item{dof}{An integer denoting the degrees of freedom of the regression model used to compute
the residuals. Only required if computing adjusted r-squared.}
}
\value{
A list with the following items:
\itemize{
\item r2: A numeric value corresponding to the unadjusted r-squared.
\item adjusted_r2: A numeric value (or NA) corresponding to the adjusted r-squared.
}
}
\description{
Calculates regression model r-squared across a matrix of outcomes.
}
\seealso{
\code{\link{get_residual_matrix}}
}
