% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{get_residual_matrix}
\alias{get_residual_matrix}
\title{Calculate residual matrix}
\usage{
get_residual_matrix(design, outcomes)
}
\arguments{
\item{design}{A data frame with explanatory variables.}

\item{outcomes}{A data frame like object with outcome variables as rows and observations as
columns.}
}
\value{
A list with the following items:
\itemize{
\item residuals: A numeric matrix with each column representing the residuals for an outcome. The
ordering follows the row order of \code{outcomes}.
\item dof: Model degrees of freedom.
}
}
\description{
Get residuals from fitting a linear model with a matrix of outcomes.
}
\details{
Calculates regression residuals using a fixed set of explanatory variables and a matrix of
outcomes.
}
\seealso{
\code{\link{get_matrix_r2}} \code{\link{get_design_mat}}
}
