% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{ggscatter}
\alias{ggscatter}
\title{ggplot2-based scatterplot}
\usage{
ggscatter(
  data,
  x_colname,
  y_colname,
  group_var = NULL,
  title = NULL,
  subtitle = NULL,
  x_title = NULL,
  y_title = NULL,
  point_size = 2,
  point_shape = 21,
  point_alpha = 0.75,
  point_color = "white",
  point_fill = "gray30",
  palette = "Paired",
  equal_axes = F,
  diag = F
)
}
\arguments{
\item{data}{A data frame or matrix with columns of interest for plotting.}

\item{x_colname}{A string for the column variable to assign to the x-axis of the plot.}

\item{y_colname}{A string for the column variable to assign to the y-axis of the plot.}

\item{group_var}{A string or vector of strings for the grouping variable(s) to use from
\code{data} to color points. If multiple variables are specified, they are combined into a single
string to make a single new variable.}

\item{title}{A string for the plot title.}

\item{subtitle}{A string for the plot subtitle.}

\item{x_title}{A string for the x-axis title.}

\item{y_title}{A string for the y-axis title.}

\item{point_size}{A numeric for the plot point size.}

\item{point_shape}{A numeric for the plot point shape.}

\item{point_alpha}{A numeric for the plot point alpha level.}

\item{point_color}{A string for plot point color. Only used when \code{group_var} is \code{NULL}.}

\item{point_fill}{A string for the plot point fill color. Only used when \code{group_var} is \code{NULL}.}

\item{palette}{A string for the RColorBrewer palette name to use when \code{group_var} is specified.}

\item{equal_axes}{A logical. Should the scatterplot use the same axis limits for both axes?}

\item{diag}{A boolean denoting whether to include a 45-degree line through the origin.}
}
\value{
A ggplot2 object.
}
\description{
ggplot2-based scatterplot
}
