% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{grouped_barplot}
\alias{grouped_barplot}
\title{Grouped barplot}
\usage{
grouped_barplot(
  data,
  x_var,
  y_var,
  rm_ids = NULL,
  fill = NULL,
  alpha = 1,
  lwd = 1,
  line_color = "black",
  x_title = NULL,
  y_title = NULL,
  legend_title = NULL,
  use_na = c("ifany", "never", "always"),
  use_proportions = F
)
}
\arguments{
\item{data}{A data frame with columns from which to retrieve variables to plot.}

\item{x_var}{A string denoting the column name of the primary barplot variable.}

\item{y_var}{A string denoting the column name of the secondary barplot variable.}

\item{rm_ids}{A string vector of the row names to exclude prior to plotting.}

\item{fill}{A string denoting the line color for barplot bars.}

\item{alpha}{A numeric value for the alpha level for barplot bars.}

\item{lwd}{A numeric value for the barplot bar line width.}

\item{x_title}{A string denoting the x-axis title.}

\item{y_title}{A string denoting the y-axis title.}

\item{legend_title}{A string denoting the legend title. Should correspond to \code{y_var}.}

\item{use_na}{One of "never", "ifany", or "always" to denote whether NA values should be
included in the frequency calculations.}

\item{use_proportions}{A boolean denoting if proportions should be used instead of frequencies.}
}
\value{
A ggplot object.
}
\description{
Creates a grouped barplot.
}
\details{
Creates a barplot for frequencies of two factor variables in which one variable is used to
stratify the other.
}
