% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{hist_boxplot}
\alias{hist_boxplot}
\title{Histogram and jittered boxplot}
\usage{
hist_boxplot(
  data,
  binsize = diff(range(data[, 1]))/50,
  hist_fill = "gray10",
  hist_alpha = 0.75,
  box_fill = "goldenrod",
  box_alpha = 0.5,
  box_lwd = 1,
  jitter_color = "gray30",
  jitter_alpha = 0.75,
  jitter_size = 1.75,
  x_title = "",
  y_title = ""
)
}
\arguments{
\item{data}{A single column data frame with numeric values to be plotted.}

\item{binsize}{A numeric value for the histogram bin widths.}

\item{hist_fill}{A string. The fill color for histogram bars.}

\item{hist_alpha}{A numeric value for the alpha level for histogram bars.}

\item{box_fill}{A string. The fill color for the boxplot.}

\item{box_alpha}{A numeric value for the alpha level for the boxplot.}

\item{box_lwd}{A numeric value for the boxplot line width.}

\item{jitter_color}{A string. The point color for jittered data points.}

\item{jitter_alpha}{A numeric for the alpha level for jittered data points.}

\item{jitter_size}{A numeric for the point size for jittered data points.}

\item{x_title}{A string denoting the x-axis title. Only added to the boxplot.}

\item{y_title}{A string denoting the y-axis title. Only added to the histogram.}
}
\value{
A list of two ggplot objects \code{hist} and \code{boxplot}.
}
\description{
Creates a histogram and jittered boxplot.
}
\details{
Creates two plots that can be organized as a single column two row multi-figure. The top row
plot is a histogram. The bottom row plot is a horizontal boxplot with jittered data points
overlayed. The x-axis of both plots are fixed to the same limits for comparability across plots.
}
