% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{hist_boxplot2}
\alias{hist_boxplot2}
\title{Histogram and jittered boxplots partitioned by categories}
\usage{
hist_boxplot2(
  data,
  binsize = diff(range(data[, 1]))/25,
  colors = NULL,
  hist_alpha = 0.75,
  box_fill = "gray50",
  box_alpha = 0.5,
  box_lwd = 1,
  jitter_alpha = 0.75,
  jitter_size = 1.75,
  x_title = "",
  y_title = ""
)
}
\arguments{
\item{data}{A two-column data frame with numeric values for plotting in the first column and
the categorical variable as the second column.}

\item{binsize}{A numeric value for the histogram bin widths.}

\item{colors}{A string vector. The histogram bar colors and jitter colors for each categorical
variable value.}

\item{hist_alpha}{A numeric value for the alpha level for histogram bars.}

\item{box_fill}{A string. The fill color for the boxplot.}

\item{box_alpha}{A numeric value for the alpha level for the boxplot.}

\item{box_lwd}{A numeric value for the boxplot line width.}

\item{jitter_alpha}{A numeric for the alpha level for jittered data points.}

\item{jitter_size}{A numeric for the point size for jittered data points.}

\item{x_title}{A string denoting the x-axis title. Only added to the boxplot.}

\item{y_title}{A string denoting the y-axis title. Only added to the histogram.}
}
\value{
A list of two ggplot objects \code{hist} and \code{boxplot}.
}
\description{
Creates a histogram and jittered boxplot with color-coded, separate distributions for each value
of a user-specified categorical variable.
}
\details{
Creates two plots that can be organized as a single column two row multi-figure. The top row
plot is multiple overlayed histograms from data partitioned by a variable. The bottom row plot
is multiple horizontal boxplots with jittered data points overlayed. The x-axis of both plots
are fixed to the same limits for comparability across plots.
}
\seealso{
\code{\link{hist_boxplot}}
}
