% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw.R
\name{ihw_diagnostic_plots}
\alias{ihw_diagnostic_plots}
\title{IHW diagnostic plots}
\usage{
ihw_diagnostic_plots(ihw_data, num_strata = nlevels(groups_factor(ihw_data)))
}
\arguments{
\item{ihw_data}{An ihwResult object.}
}
\value{
A list of ggplot objects.
}
\description{
Creates ggplot objects from an ihwResult object. The ggplot objects correspond to eight plots
that assess if IHW is a reasonable method to use with the user-specified independent variable:
\itemize{
\item log p-value vs. independent variable ranking
\item log p-value density plots stratified by independent variable
\item p-value histograms stratified by independent variable
\item p-value cumulative distributions stratified by independent variable
\item mean IHW weight stratified by independent variable for multiple folds of the data
\item decision boundary relative to p-values and independent variable ranking
\item IHW adjusted p-values vs. raw p-values
\item IHW adjusted p-values vs. adjusted p-values using a different method
This function is a convenient wrapper to quickly generate several diagnostic plots, but each
plot can be generated separately using its respective function.
}
}
\details{
Create diagnostic plots for independent hypothesis weighting.
}
\seealso{
\code{\link{ihw_pval_scatter}}, \code{\link{ihw_pval_ridges}},
\code{\link{ihw_stratified_pval_hist}}, \code{\link{ihw_pval_cdf}},
\code{\link{ihw_partitioned_pval_weights}}, \code{\link{ihw_pval_comparison}},
\code{\link{ihw_pval_decision_boundary}},

Other ihw: 
\code{\link{ihw_partitioned_pval_weights}()},
\code{\link{ihw_pval_cdf}()},
\code{\link{ihw_pval_comparison}()},
\code{\link{ihw_pval_decision_boundary}()},
\code{\link{ihw_pval_ridges}()},
\code{\link{ihw_pval_scatter}()},
\code{\link{ihw_stratified_pval_hist}()}
}
\concept{ihw}
