% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw.R
\name{ihw_partitioned_pval_weights}
\alias{ihw_partitioned_pval_weights}
\title{Stratified IHW p-value weights}
\usage{
ihw_partitioned_pval_weights(
  ihw_data,
  scale = covariate_type(ihw_data),
  point_size = 2,
  line_size = 1.5,
  alpha = 1
)
}
\arguments{
\item{ihw_data}{An ihwResult object.}

\item{scale}{Scale of the IHW variable. Either 'nominal' or 'ordinal'.}

\item{point_size}{Point size for ordinal plot.}

\item{line_size}{Line size for ordinal plot.}

\item{alpha}{Plot colors alpha value.}
}
\value{
A ggplot object.
}
\description{
Plot p-value weights for each strata partitioned by fold.
}
\details{
Produces either a barplot or line plot depending on if the IHW variable is nominal or ordinal.
A stable p-value weighting scheme is expected to show consistent weights across folds within a
stratum and increasing weights with increasing stratum ID.
}
\seealso{
Other ihw: 
\code{\link{ihw_diagnostic_plots}()},
\code{\link{ihw_pval_cdf}()},
\code{\link{ihw_pval_comparison}()},
\code{\link{ihw_pval_decision_boundary}()},
\code{\link{ihw_pval_ridges}()},
\code{\link{ihw_pval_scatter}()},
\code{\link{ihw_stratified_pval_hist}()}
}
\concept{ihw}
