% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw.R
\name{ihw_pval_cdf}
\alias{ihw_pval_cdf}
\title{Stratified p-value eCDFs}
\usage{
ihw_pval_cdf(ihw_data, num_strata = nlevels(groups_factor(ihw_data)))
}
\arguments{
\item{ihw_data}{An ihwResult object.}

\item{num_strata}{Number of strata into which the independent variable is divided.}
}
\value{
A ggplot object.
}
\description{
Plot empirical CDFs for p-values stratified by the IHW variable.
}
\details{
This plot partitions p-values into multiple eCDFs by IHW variable strata. An informative
IHW variable will show a strongly increasing area under curve as the stratum ID increases.
}
\seealso{
Other ihw: 
\code{\link{ihw_diagnostic_plots}()},
\code{\link{ihw_partitioned_pval_weights}()},
\code{\link{ihw_pval_comparison}()},
\code{\link{ihw_pval_decision_boundary}()},
\code{\link{ihw_pval_ridges}()},
\code{\link{ihw_pval_scatter}()},
\code{\link{ihw_stratified_pval_hist}()}
}
\concept{ihw}
