% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw.R
\name{ihw_pval_comparison}
\alias{ihw_pval_comparison}
\title{Compare p-value transformations}
\usage{
ihw_pval_comparison(ihw_data, p_adjust = NULL, point_size = 1.5, alpha = 1)
}
\arguments{
\item{ihw_data}{An ihwResult object.}

\item{p_adjust}{A p-value adjustment compatible with the 'method' parameter in 'p.adjust'. If
'NULL' then no p-value adjustment is made.}

\item{point_size}{Point size for plot.}

\item{alpha}{Plot colors alpha value.}
}
\value{
A ggplot object.
}
\description{
Plot IHW adjusted p-values to raw p-values or adjusted p-values using other methods.
}
\details{
Creates a scatterplot that compares IHW adjusted p-values to p-values adjusted using a
different method (or simply the unadjusted p-values).
}
\seealso{
Other ihw: 
\code{\link{ihw_diagnostic_plots}()},
\code{\link{ihw_partitioned_pval_weights}()},
\code{\link{ihw_pval_cdf}()},
\code{\link{ihw_pval_decision_boundary}()},
\code{\link{ihw_pval_ridges}()},
\code{\link{ihw_pval_scatter}()},
\code{\link{ihw_stratified_pval_hist}()}
}
\concept{ihw}
