% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw.R
\name{ihw_pval_decision_boundary}
\alias{ihw_pval_decision_boundary}
\title{Plot IHW weight decision boundaries}
\usage{
ihw_pval_decision_boundary(
  ihw_data,
  scale = covariate_type(ihw_data),
  line_size = 1.5,
  alpha = 1
)
}
\arguments{
\item{ihw_data}{An ihwResult object.}

\item{scale}{Scale of the IHW variable. Either 'nominal' or 'ordinal'.}

\item{line_size}{Line size for decision boundary.}

\item{alpha}{Plot colors alpha value.}
}
\value{
A ggplot object.
}
\description{
Plot p-value weight decision boundaries by strata partitioned by fold.
}
\seealso{
Other ihw: 
\code{\link{ihw_diagnostic_plots}()},
\code{\link{ihw_partitioned_pval_weights}()},
\code{\link{ihw_pval_cdf}()},
\code{\link{ihw_pval_comparison}()},
\code{\link{ihw_pval_ridges}()},
\code{\link{ihw_pval_scatter}()},
\code{\link{ihw_stratified_pval_hist}()}
}
\concept{ihw}
