% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw.R
\name{ihw_pval_scatter}
\alias{ihw_pval_scatter}
\title{Plot p-value vs. IHW variable ranking}
\usage{
ihw_pval_scatter(
  ihw_data,
  truncate_y = F,
  point_color = "gray20",
  point_alpha = 0.5,
  point_size = 2
)
}
\arguments{
\item{ihw_data}{An ihwResult object.}

\item{truncate_y}{If 'TRUE', remove -log10 p-value outliers from the plot (exceed 1.5*IQR above
the third quartile).}

\item{point_color}{Scatterplot point fill color.}

\item{point_alpha}{Scatterplot point fill color alpha value.}

\item{point_size}{Scatterplot point size.}
}
\value{
A ggplot object.
}
\description{
Plots feature p-values vs. feature IHW variable rankings.
}
\details{
This plot provides insight into how informative the selected IHW variable is in relation to
the hypothesis being tested (e.g., differential gene expression). As IHW variable ranking
increases (i.e., the rank gets larger) the p-values should trend smaller for more informative
variables.
}
\seealso{
Other ihw: 
\code{\link{ihw_diagnostic_plots}()},
\code{\link{ihw_partitioned_pval_weights}()},
\code{\link{ihw_pval_cdf}()},
\code{\link{ihw_pval_comparison}()},
\code{\link{ihw_pval_decision_boundary}()},
\code{\link{ihw_pval_ridges}()},
\code{\link{ihw_stratified_pval_hist}()}
}
\concept{ihw}
