% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw.R
\name{ihw_stratified_pval_hist}
\alias{ihw_stratified_pval_hist}
\title{Stratified p-value histograms}
\usage{
ihw_stratified_pval_hist(
  ihw_data,
  num_strata = nlevels(groups_factor(ihw_data)),
  fill_color = "gray20",
  fill_alpha = 1
)
}
\arguments{
\item{ihw_data}{An ihwResult object.}

\item{num_strata}{Number of strata into which the independent variable is divided.}

\item{fill_color}{Fill color for area under density distributions.}

\item{fill_alpha}{Fill color alpha value for area under density distributions.}
}
\value{
A ggplot object.
}
\description{
Plot p-value histograms for IHW variable strata.
}
\details{
This plot partitions the p-values into equal sized strata of the IHW variable to generate
multiple p-value histograms. For an informative IHW variable, the larger the stratum ID the
more skew there should be towards enrichment of smaller p-values.
}
\seealso{
Other ihw: 
\code{\link{ihw_diagnostic_plots}()},
\code{\link{ihw_partitioned_pval_weights}()},
\code{\link{ihw_pval_cdf}()},
\code{\link{ihw_pval_comparison}()},
\code{\link{ihw_pval_decision_boundary}()},
\code{\link{ihw_pval_ridges}()},
\code{\link{ihw_pval_scatter}()}
}
\concept{ihw}
