% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R, R/dge_utils.R, R/limma.R
\name{low_expression_filter.DESeqDataSet}
\alias{low_expression_filter.DESeqDataSet}
\alias{low_expression_filter}
\alias{low_expression_filter.matrix}
\alias{low_expression_filter.DGEList}
\title{Filter lowly expressed features prior to DGE analysis}
\usage{
\method{low_expression_filter}{DESeqDataSet}(
  dds,
  value_cutoff,
  min_sample_fraction = NULL,
  threshold_variable = NULL,
  use_min_fraction = T,
  approximate = F
)

low_expression_filter(
  object,
  value_cutoff,
  min_sample_fraction = NULL,
  threshold_variable = NULL,
  use_min_fraction = T,
  approximate = F
)

\method{low_expression_filter}{matrix}(object, value_cutoff, sample_cutoff, approximate)

\method{low_expression_filter}{DGEList}(
  dge_data,
  value_cutoff,
  min_sample_fraction = NULL,
  threshold_variable = NULL,
  use_min_fraction = T,
  approximate = F
)
}
\arguments{
\item{value_cutoff}{The minimum value required for a feature for greater than
\code{min_sample_fraction} proportion of samples to retain the feature.}

\item{min_sample_fraction}{The minimum proportion of samples with greater than \code{value_cutoff}
values to retain a feature.}

\item{threshold_variable}{Name of the binary variable to use for calculating sample fractions.}

\item{use_min_fraction}{If \code{TRUE}, uses the smaller of the sample fractions calculated from the
\code{threshold_variable}. Otherwise uses the larger of the sample fractions.}

\item{approximate}{If \code{TRUE}, round \code{min_sample_fraction} to a number with the hundreths digit
equal to '0' or '5'.}

\item{object}{A DESeqDataSet, DGEList, or matrix object from which to filter row features.}
}
\value{
A filtered DESeqDataSet object if a DESeqDataSet object is provided as input.

A filtered matrix if a matrix is provided as input.

A filtered DGEList object if a DGEList object is provided as input.
}
\description{
Generic method to remove expression features below a given threshold.
}
\details{
Filters out features considered lowly expressed using a count threshold and sample proportion
set by the user. The sample proportion can also be estimated by calculating proportions of
values from a binary variable specified by the user.
}
