% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode_utils.R
\name{make_gencode_gene_id_map}
\alias{make_gencode_gene_id_map}
\title{Create GENCODE ID to gene name mapping}
\usage{
make_gencode_gene_id_map(gtf_file, output_file = NULL)
}
\arguments{
\item{gtf_file}{A string denoting a GTF file from which to build the mapping.}

\item{output_file}{Optional path and name of file to which to save the mapping table.}
}
\value{
A data frame with each row linking a GENCODE ID to a gene name and HGNC ID.
}
\description{
Creates a GENCODE ID to gene name mapping using a GTF file.
}
