% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_utils.R
\name{merge_txi}
\alias{merge_txi}
\title{Merge tximport objects}
\usage{
merge_txi(txi_list)
}
\arguments{
\item{txi_list}{A list where each element is a tximport object.}
}
\value{
A list with the following components:
\itemize{
\item \code{txi} A tximport object.
\item \code{sample_source} A categorical variable for each sample that corresponds to the source
tximport object for the sample.
}
}
\description{
Combines multiple tximport objects into a single tximport object.
}
\details{
Merges tximport objects to create a single tximport object with a common subset of genes across
all of the input objects. Also creates a categorical variable to distinguish the source object
of each sample. Requires all row and sample names to be unique.
}
