% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow_qc.R
\name{parse_by_trim_status}
\alias{parse_by_trim_status}
\title{Partition rows by trimming status.}
\usage{
parse_by_trim_status(
  data,
  str_query = "trimmed",
  paired_end = F,
  r1_query = "R1",
  r2_query = "R2",
  unpaired_query = "unpaired"
)
}
\arguments{
\item{data}{A data frame with rows corresponding to samples with names that differentiate
different stages of sample processing from the RNA-seq data processing workflow (pre vs.
post-Trimmomatic and "R1" and "R2" for read pairs in paired-end sequencing).}

\item{str_query}{A string in the row names for \code{data} that distinguish pre-Trimmomatic from
post-Trimmomatic records.}

\item{paired_end}{A logical. Is the data from paired-end sequencing? If so, additional
partitions will be made that separate the forward and reverse read pairs.}

\item{r1_query}{A string in the row names for \code{data} that distinguish R1 reads from others in
paired-end sequencing data.}

\item{r2_query}{A string in the row names for \code{data} that distinguish R2 reads from others in
paired-end sequencing data.}

\item{unpaired_query}{A string in the row names for \code{data} that distinguish unpaired reads from
preserved pairs in paired-end sequencing data.}
}
\value{
For single-end sequencing data, a list with the elements \code{untrimmed} and \code{trimmed} to
denote data frames with pre-Trimmomatic and post-Trimmomatic data respectively. For
paired-end sequencing data, a list with the following elements:
\itemize{
\item \code{untrimmed_r1}: Data frame with pre-Trimmomatic R1 reads.
\item \code{untrimmed_r2}: Data frame with pre-Trimmomatic R2 reads.
\item \code{trimmed_r1}: Data frame with post-Trimmomatic R1 reads from preserved pairs.
\item \code{trimmed_r2}: Data frame with post-Trimmomatic R2 reads from preserved pairs.
\item \code{unpaired_r1}: Data frame with post-Trimmomatic R1 reads without a read pair.
\item \code{unpaired_r2}: Data frame with post-Trimmomatic R2 reads without a read pair.
}
}
\description{
Partition rows of a data frame by trimming status using string matching on row names.
}
\details{
Parses a quality metric table that has data from multiple stages of the RNA-seq data processing
workflow into separate tables for pre-Trimmomatic and post-Trimmomatic data. If the data are
from paired-end sequencing, additional partitions can be made that separate post-Trimmomatic
preserved read pairs from post-Trimmomatic unpaired reads, and read 1 (R1) and read 2 (R2). This
function is intended for use with auto-generated QC metrics tables from the RNA-seq data
processing workflow and tables exported from MultiQC HTML reports.
}
