% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R, R/dge_utils.R, R/limma.R
\name{per_sample_count_distribution.DESeqDataSet}
\alias{per_sample_count_distribution.DESeqDataSet}
\alias{per_sample_count_distribution}
\alias{per_sample_count_distribution.matrix}
\alias{per_sample_count_distribution.DGEList}
\title{Plot per sample count distribution}
\usage{
\method{per_sample_count_distribution}{DESeqDataSet}(
  dds,
  normalized = T,
  point_size = 2.5,
  point_alpha = 1,
  y_lim = NULL
)

per_sample_count_distribution(object, ...)

\method{per_sample_count_distribution}{matrix}(
  data,
  point_size = 2.5,
  point_alpha = 1,
  y_lim = NULL
)

\method{per_sample_count_distribution}{DGEList}(
  dge_data,
  normalized = T,
  point_size = 2.5,
  point_alpha = 1,
  y_lim = NULL
)
}
\arguments{
\item{normalized}{A boolean denoting whether the plotted counts should not be normalized
(\code{FALSE}) or normalized by size/normalization factors (\code{TRUE}).
factors.}

\item{point_size}{Plot point size.}

\item{point_alpha}{Plot point alpha value.}

\item{y_lim}{Plot y-axis range.}

\item{object}{A DESeqDataSet, DGEList, or numeric matrix object.}
}
\value{
A ggplot object.
}
\description{
Plot feature count distribution summary statistics per sample.
}
\details{
Plots the 25th, 50th, and 75th percentil of the feature count distribution for each sample.
}
