% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{plot_cooks_per_sample}
\alias{plot_cooks_per_sample}
\title{Plot Cook's distance distributions per sample.}
\usage{
plot_cooks_per_sample(
  dds_fit,
  point_size = 2.5,
  point_alpha = 1,
  cooks_cutoff = NULL,
  y_lim = NULL
)
}
\arguments{
\item{dds_fit}{A DESeqDataSet object after model fitting.}

\item{point_alpha}{Plot point color alpha value.}

\item{cooks_cutoff}{Cook's distance cutoff value for calling outliers.}

\item{y_lim}{y-axis lower and upper plot limit.}

\item{point_color}{Plot point color.}
}
\value{
A ggplot object.
}
\description{
Visualizes Cook's distance distribution summary stats across all features for a sample.
}
\details{
Uses Cook's distance values across all features and samples to create a plot with the 25th,
50th, and 75th percentiles of the Cook's distance distrubtion for each sample.
}
\seealso{
\code{\link{get_cooks_cutoff}}
}
