% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{plot_cooks_vs_wald}
\alias{plot_cooks_vs_wald}
\title{Compare Cook's distances to Wald statistic ranks}
\usage{
plot_cooks_vs_wald(
  dds_fit,
  dds_results,
  f_threshold = 99,
  point_size = 3,
  point_color = "gray30",
  point_alpha = 0.5
)
}
\arguments{
\item{dds_fit}{A DESeqDataSet object after model fitting.}

\item{dds_results}{A DESeqResults object derived from \code{dds_fit}.}

\item{f_threshold}{Percentile of the F distribution used as the cutoff for calling outliers.}

\item{point_size}{Plot point size.}

\item{point_color}{Plot point color.}

\item{point_alpha}{Plot point color alpha value.}
}
\value{
A ggplot object.
}
\description{
Plot max Cook's distances per feature vs. Wald statistic rankings.
}
\details{
Obtains the max Cook's distance for a feature across all samples and compares it to the Wald
statistic derived from the statistical test for differential expression. Extreme Cook's
distance values are truncated to the y-axis upper limit to aid with visualization.
}
