% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow_qc.R
\name{plot_decoy_mapping_rate}
\alias{plot_decoy_mapping_rate}
\title{Plot Salmon percent decoy fragments}
\usage{
plot_decoy_mapping_rate(
  data,
  ids = NULL,
  invert = F,
  binsize = NULL,
  hist_fill = "gray10",
  hist_alpha = 0.75,
  box_fill = "goldenrod",
  box_alpha = 0.5,
  box_lwd = 1,
  jitter_color = "gray30",
  jitter_alpha = 0.75,
  jitter_size = 1.75,
  x_title = "\% mapped to decoys",
  y_title = "Sample count"
)
}
\arguments{
\item{data}{A data frame with Salmon mapping data in the same format as \code{salmon} returned by
\code{\link{load_paired_end_qc_data}}.}

\item{ids}{A vector of rownames for subsetting \code{data} for plotting.}

\item{invert}{A logical. Should \code{ids} be used for excluding rows from plotting instead?}

\item{binsize}{A numeric value for the histogram bin widths.}

\item{hist_alpha}{A numeric value for the alpha level for histogram bars.}

\item{box_fill}{A string. The fill color for the boxplot.}

\item{box_alpha}{A numeric value for the alpha level for the boxplot.}

\item{box_lwd}{A numeric value for the boxplot line width.}

\item{jitter_alpha}{A numeric for the alpha level for jittered data points.}

\item{jitter_size}{A numeric for the point size for jittered data points.}

\item{x_title}{A string denoting the x-axis title. Only added to the boxplot.}

\item{y_title}{A string denoting the y-axis title. Only added to the histogram.}
}
\value{
A list with two ggplot objects \code{hist} and \code{boxplot} corresponding to a histogram and
jittered boxplot respectively.
}
\description{
Create a Salmon decoy mapping rate histogram and boxplot. The two plots are intended to be used
as a single column, two row figure panel.
}
\seealso{
\code{\link{load_paired_end_qc_data}}
}
