% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow_qc.R
\name{plot_expression_fraction}
\alias{plot_expression_fraction}
\title{Plot feature set expression fraction}
\usage{
plot_expression_fraction(
  data,
  group = NULL,
  feature_ids = NULL,
  sample_ids = NULL,
  invert_rows = F,
  invert_cols = F,
  return_data = F,
  binsize = NULL,
  colors = NULL,
  hist_alpha = 0.75,
  box_fill = "goldenrod",
  box_alpha = 0.5,
  box_lwd = 1,
  jitter_alpha = 0.75,
  jitter_size = 1.75,
  x_title = "\% of total expression",
  y_title = "Sample count"
)
}
\arguments{
\item{data}{A feature (row) by sample (column) data frame or matrix with expression count
values.}

\item{group}{A single column data frame denoting categorical variable values to use to
partition the data points into separate distributions/boxplots. Row names should match
rownames in \code{data}.}

\item{feature_ids}{A vector of row names for subsetting \code{data} for plotting.}

\item{sample_ids}{A vector of column names for subsetting \code{data} for plotting.}

\item{invert_rows}{A logical. Should \code{feature_ids} be used for excluding rows from plotting
instead?}

\item{invert_cols}{A logical. Should \code{sample_ids} be used for excluding columns from plotting
instead?}

\item{return_data}{A logical. Should plot data be returned instead of a ggplot object?}

\item{binsize}{A numeric value for the histogram bin widths.}

\item{colors}{A string vector. The histogram bar colors and jitter colors for each categorical
variable value.}

\item{hist_alpha}{A numeric value for the alpha level for histogram bars.}

\item{box_fill}{A string. The fill color for the boxplot.}

\item{box_alpha}{A numeric value for the alpha level for the boxplot.}

\item{box_lwd}{A numeric value for the boxplot line width.}

\item{jitter_alpha}{A numeric for the alpha level for jittered data points.}

\item{jitter_size}{A numeric for the point size for jittered data points.}

\item{x_title}{A string denoting the x-axis title. Only added to the boxplot.}

\item{y_title}{A string denoting the y-axis title. Only added to the histogram.}
}
\value{
A list with two ggplot objects \code{hist} and \code{boxplot} corresponding to a histogram and
jittered boxplot respectively. If \code{return_data} is \code{TRUE}, then a data frame with the plot
data values and grouping variable.
}
\description{
Create a feature set fraction histogram and boxplot.
}
\details{
Calculates the fraction of feature (e.g., gene) expression represented by a user-specified set
of features and plots this as a histogram and boxplot. The two plots are intended to be used as
a single column, two row figure panel.
}
\seealso{
\verb{\link{load_paired_end_qc_data}} \verb{\link{hist_boxplot2}}
}
