% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow_qc.R
\name{plot_frag_length}
\alias{plot_frag_length}
\title{Fragment length frequency line plot}
\usage{
plot_frag_length(
  data,
  line_width = 1.5,
  line_color = "red4",
  alpha = 0.5,
  ids = NULL,
  invert = F
)
}
\arguments{
\item{data}{A data frame with fragment length data in the same format as \code{frag_length} returned
by \code{\link{load_paired_end_qc_data}}.}

\item{line_width}{A numeric. The line size for plot lines.}

\item{line_color}{A string. The color for plot lines.}

\item{alpha}{A numeric. The alpha level for data points.}

\item{ids}{A vector of rownames for subsetting \code{data} for plotting.}

\item{invert}{A logical. Should \code{ids} be used for excluding rows from plotting instead?}
}
\value{
A ggplot object.
}
\description{
Create a fragment length frequency line plot.
}
\seealso{
\code{\link{load_paired_end_qc_data}}
}
