% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow_qc.R
\name{plot_gc_content}
\alias{plot_gc_content}
\title{GC content distribution line plot}
\usage{
plot_gc_content(
  data,
  labels = NULL,
  line_colors = rep("red4", length(data)),
  alpha = 0.25,
  line_width = 1.5,
  ids = NULL,
  invert = F
)
}
\arguments{
\item{data}{A list of data frames with data in the same format as \code{gc_content} returned
by \verb{\link{load_paired_end_qc_data}}. Each data frame is intended to represent a different
category of data points (e.g., R1 vs. R2 reads in paired-end data).}

\item{labels}{A vector of strings for legend labels corresponding to the data frames in \code{data}.}

\item{line_colors}{A vector of strings. Line colors corresponding to the data frames in \code{data}.}

\item{alpha}{A numeric. The alpha level for lines.}

\item{line_width}{A numeric value for the line width.}

\item{ids}{A vector of rownames for subsetting \code{data} for plotting.}

\item{invert}{A logical. Should \code{ids} be used for excluding rows from plotting instead?}
}
\value{
A ggplot object.
}
\description{
Create a line plot with read GC content distributions.
}
\seealso{
\verb{\link{load_paired_end_qc_data}}
}
