% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow_qc.R
\name{plot_gc_content_hist}
\alias{plot_gc_content_hist}
\title{GC content histogram}
\usage{
plot_gc_content_hist(
  data,
  binsize = 2.5,
  labels = NULL,
  fill = rep("red4", length(data)),
  alpha = 0.25,
  ids = NULL,
  invert = F
)
}
\arguments{
\item{data}{A list of data frames with data in the same format as \code{fastqc} returned by
\code{\link{load_paired_end_qc_data}}.}

\item{binsize}{A numeric value for the histogram bin widths.}

\item{labels}{A vector of strings for legend labels corresponding to the data frames in \code{data}.}

\item{fill}{A string. The fill color for histogram bars.}

\item{alpha}{A numeric. The alpha level for histogram bars.}

\item{ids}{A vector of rownames for subsetting \code{data} for plotting.}

\item{invert}{A logical. Should \code{ids} be used for excluding rows from plotting instead?}
}
\value{
A ggplot object.
}
\description{
Create a GC content by sample histogram.
}
\seealso{
\code{\link{load_paired_end_qc_data}}
}
