% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow_qc.R
\name{plot_hisat2_vs_salmon}
\alias{plot_hisat2_vs_salmon}
\title{Plot HISAT2 vs. Salmon mapping rates}
\usage{
plot_hisat2_vs_salmon(
  hisat_data,
  salmon_data,
  point_size = 4,
  fill = "red4",
  alpha = 0.75,
  ids = NULL,
  invert = F
)
}
\arguments{
\item{hisat_data}{A data frame of HISAT2 mapping data in the same format as \code{hisat2} returned by
\verb{\link{load_paired_end_qc_data}}.}

\item{salmon_data}{A data frame of Salmon mapping data in the same format as \code{hisat2} returned
by \verb{\link{load_paired_end_qc_data}}.}

\item{point_size}{A numeric. The size for data points.}

\item{fill}{A string. The fill color for data points.}

\item{alpha}{A numeric. The alpha level for data points.}

\item{ids}{A vector of rownames for subsetting \code{hisat_data} and \code{salmon_data} for plotting.}

\item{invert}{A logical. Should \code{ids} be used for excluding rows from plotting instead?}
}
\value{
A ggplot object.
}
\description{
Create a HISAT2 vs Salmon mapping rate scatterplot. Only applicable for
paired-end sequencing data.
}
\details{
Creates a scatterplot comparing mapping rates between HISAT2 and Salmon for samples that have
matchable IDs between data from the two tools. IDs are formatted  for matching based on the
assumption that sample names are equivalent up until the first \code{|} delimiter (if any) in the
name.
}
\seealso{
\verb{\link{load_paired_end_qc_data}}
}
