% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow_qc.R
\name{plot_inner_dist}
\alias{plot_inner_dist}
\title{Read pair inner distance distribution}
\usage{
plot_inner_dist(
  data,
  line_width = 1.5,
  line_color = "red4",
  alpha = 0.5,
  ids = NULL,
  invert = F
)
}
\arguments{
\item{data}{A data frame with fragment length data in the same format as \code{inner_dist} returned
by \verb{\link{load_paired_end_qc_data}}.}

\item{line_width}{A numeric. The line size for plot lines.}

\item{line_color}{A string. The color for plot lines.}

\item{alpha}{A numeric. The alpha level for data points.}

\item{ids}{A vector of rownames for subsetting \code{data} for plotting.}

\item{invert}{A logical. Should \code{ids} be used for excluding rows from plotting instead?}
}
\value{
A ggplot object.
}
\description{
Create a paired-end read inner distance line plot. Applicable only to paired-end sequencing
data.
}
\seealso{
\verb{\link{load_paired_end_qc_data}}
}
