% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow_qc.R
\name{plot_partitioned_mean_expression}
\alias{plot_partitioned_mean_expression}
\title{Plot partitioned mean expression}
\usage{
plot_partitioned_mean_expression(
  data,
  group_var,
  feature_ids = NULL,
  sample_ids = NULL,
  invert_rows = F,
  invert_cols = F,
  return_data = F,
  binsize = NULL,
  colors = NULL,
  hist_alpha = 0.75,
  box_fill = "gray50",
  box_alpha = 0.5,
  box_lwd = 1,
  jitter_alpha = 0.5,
  jitter_size = 1.75,
  x_title = "Mean expression",
  y_title = "Sample count"
)
}
\arguments{
\item{data}{A SummarizedExperiment-like object. Must be compatible with \code{assay()} and
\code{colData()}.}

\item{group_var}{A string for the categorical variable to use for partitioning.}

\item{feature_ids}{A vector of row names for subsetting \code{data} for plotting.}

\item{sample_ids}{A vector of column names for subsetting \code{data} for plotting.}

\item{invert_rows}{A logical. Should \code{feature_ids} be used for excluding rows from plotting
instead?}

\item{invert_cols}{A logical. Should \code{sample_ids} be used for excluding columns from plotting
instead?}

\item{return_data}{A logical. Should plot data be returned instead of a ggplot object?}

\item{binsize}{A numeric value for the histogram bin widths.}

\item{colors}{A string vector. The histogram bar colors and jitter colors for each categorical
variable value.}

\item{hist_alpha}{A numeric value for the alpha level for histogram bars.}

\item{box_fill}{A string. The fill color for the boxplot.}

\item{box_alpha}{A numeric value for the alpha level for the boxplot.}

\item{box_lwd}{A numeric value for the boxplot line width.}

\item{jitter_alpha}{A numeric for the alpha level for jittered data points.}

\item{jitter_size}{A numeric for the point size for jittered data points.}

\item{x_title}{A string denoting the x-axis title. Only added to the boxplot.}

\item{y_title}{A string denoting the y-axis title. Only added to the histogram.}
}
\value{
A list with two ggplot objects \code{hist} and \code{boxplot} corresponding to a histogram and
jittered boxplot respectively. If \code{return_data} is \code{TRUE}, then a data frame with the plot
data values and grouping variable.
}
\description{
Create a histogram and boxplot with per sample mean expression levels partitioned by a
categorical variable.
}
\details{
Calculates the mean across a user-specified set of expression features (e.g., genes) for each
sample and plots this as a histogram and boxplot. The two plots are intended to be used as a
single column, two row figure panel. Data points in the plots are partitioned by a categorical
variable.
}
\seealso{
\verb{\link{hist_boxplot2}}
}
