% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{plot_r2}
\alias{plot_r2}
\title{Plot (adjusted) r-squared for multiple models}
\usage{
plot_r2(
  pve_list,
  model_names = NULL,
  line_color = "gray50",
  line_alpha = 0.5,
  line_type = "dashed",
  line_size = 1.5,
  point_color = NULL,
  palette = "Paired",
  point_alpha = 1,
  point_shape = 16,
  point_size = 5,
  label_size = 5,
  label_line_size = 0.75,
  label_line_alpha = 0.5,
  title = NULL,
  x_title = "Model complexity (no. of variables)",
  y_title = "Variance explained",
  legend_title = NULL,
  legend_position = "right",
  xlim_multiplier = 0.25,
  ylim_multiplier = 0.25
)
}
\arguments{
\item{pve_list}{A list of lists containing regression model information. Each list must have the
same element ordering, and the following elements and corresponding names are required:
\itemize{
\item \code{r2}: The coefficient of determination. Can be calculated using \code{get_matrix_r2()}. Value can
be \code{NA} if \code{adjusted_r2} is numeric.
\item \code{adjusted_r2}: The adjusted r2. Can be calculated using \code{get_matrix_r2()}. Value can be \code{NA}
if \code{r2} is numeric.
\item \code{num_model_vars}: The number of variables in the model used to calculate \code{r2} and
\code{adjusted_r2}.
}}

\item{model_names}{A vector of model names that follows the same ordering as the lists in
\code{pve_list}. If not provided, a sequential numbering will be used.}

\item{line_color}{A string for the line color between points.}

\item{line_alpha}{A numeric for the line alpha value.}

\item{line_type}{A string for the line type between points.}

\item{line_size}{A numeric for the line size between points.}

\item{point_color}{A vector of colors corresponding to the groups of points (r2 vs adjusted r2).
If not specified, default colors from a color palette will be used.}

\item{palette}{A string corresponding to a color palette from which point colors are chosen.}

\item{point_alpha}{A numeric for the point alpha value.}

\item{point_shape}{A numeric for the point shape.}

\item{point_size}{A numeric for the point size.}

\item{label_size}{A numeric for the label text size..}

\item{label_line_size}{A numeric for the line thickness of the guide line from the label text
to its corresponding point.}

\item{title}{Plot title.}

\item{x_title}{X-axis title.}

\item{y_title}{Y-axis title.}

\item{legend_title}{Legend title.}

\item{xlim_multiplier}{A numeric constant used to increase or decrease the x-axis range.}

\item{ylim_multiplier}{A numeric constant used to icnrease or decrease the y-axis range.}

\item{label_line_alpha.}{A numeric for the line alpha value of the guide line from the label
text to its corresponding point.}
}
\value{
A ggplot object.
}
\description{
Plots regression model r-squared vs. model size (number of variables).
}
\details{
Creates a scatterplot showing the tradeoff between variance explained of the outcome
(e.g., gene expression) and model complexity. This function is designed to be a visualization
helper function for the output from \code{get_matrix_r2}. It allows for the plotting of only r2, only
adjusted r2, or both.
}
\seealso{
\code{\link{get_matrix_r2}} \code{\link{get_residual_matrix}}
\code{\link{get_design_mat}}
}
