% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow_qc.R
\name{plot_salmon_categories}
\alias{plot_salmon_categories}
\title{Plot Salmon mapping category percentages}
\usage{
plot_salmon_categories(
  data,
  fill = brewer.pal(n = 4, name = "Spectral"),
  sort = F,
  ids = NULL,
  invert = F
)
}
\arguments{
\item{data}{A data frame with Salmon mapping data in the same format as \code{salmon} returned by
\verb{\link{load_paired_end_qc_data}}.}

\item{fill}{A string vector with 4 colors corresponding to bars for the barplots for the
transcriptome, decoys, dovetails, and filtered mappings respectively.}

\item{sort}{A logical. Should the x-axis be ordered by transcriptome mapping rate?}

\item{ids}{A vector of rownames for subsetting \code{data} for plotting.}

\item{invert}{A logical. Should \code{ids} be used for excluding rows from plotting instead?}
}
\value{
A list with four ggplot objects \code{transcriptome}, \code{decoy}, \code{dovetail}, and \code{filtered}
that correspond to the plots for the transcriptome, decoys, dovetails, and filtered mappings
respectively.
}
\description{
Create barplots with percentages for types of Salmon mappings.
}
\details{
Create barplots with the percentage of reads/read pairs input into Salmon that were
mapped to the transcriptome, decoy fragments, as dovetails, and as low-score filtered mappings.
Categories are not mutually exclusive. X-axes are ordered the same across plots.
}
\seealso{
\verb{\link{load_paired_end_qc_data}}
}
