% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow_qc.R
\name{plot_seq_depth}
\alias{plot_seq_depth}
\title{Sequencing depth per sample scatterplot}
\usage{
plot_seq_depth(
  data,
  point_size = 4,
  fill = "red4",
  alpha = 0.75,
  sort = T,
  ids = NULL,
  invert = F
)
}
\arguments{
\item{data}{A data frame with sequencing depth data in the same format as \code{fastqc} returned
by \verb{\link{load_paired_end_qc_data}}.}

\item{point_size}{A numeric. The size for data points.}

\item{fill}{A string. The fill color for data points.}

\item{alpha}{A numeric. The alpha level for data points.}

\item{sort}{A logical. Should the x-axis be ordered by sequencing depth?}

\item{ids}{A vector of rownames for subsetting \code{data} for plotting.}

\item{invert}{A logical. Should \code{ids} be used for excluding rows from plotting instead?}
}
\value{
A ggplot object.
}
\description{
Create a sequencing depth by sample scatterplot.
}
\seealso{
\verb{\link{load_paired_end_qc_data}}
}
