% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow_qc.R
\name{plot_seq_depth_hist}
\alias{plot_seq_depth_hist}
\title{Sequencing depth histogram}
\usage{
plot_seq_depth_hist(
  data,
  binsize = NULL,
  fill = "red4",
  alpha = 0.75,
  ids = NULL,
  invert = F
)
}
\arguments{
\item{data}{A data frame with sequencing depth data in the same format as \code{fastqc} returned
by \verb{\link{load_paired_end_qc_data}}.}

\item{binsize}{A numeric value for the histogram bin widths.}

\item{fill}{A string. The fill color for histogram bars.}

\item{alpha}{A numeric. The alpha level for histogram bars.}

\item{ids}{A vector of rownames for subsetting \code{data} for plotting.}

\item{invert}{A logical. Should \code{ids} be used for excluding rows from plotting instead?}
}
\value{
A ggplot object.
}
\description{
Create a sequencing depth histogram.
}
\seealso{
\verb{\link{load_paired_end_qc_data}}
}
