% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow_qc.R
\name{plot_trimmomatic_paired}
\alias{plot_trimmomatic_paired}
\title{Trimmomatic statistics histograms}
\usage{
plot_trimmomatic_paired(
  data,
  binsize = c(2.5, 0.5),
  fill = c("gray10", "goldenrod2", "steelblue4"),
  alpha = 0.5,
  ids = NULL,
  invert = F
)
}
\arguments{
\item{data}{A data frame with sequencing depth data in the same format as \code{trimmomatic} returned
by \code{\link{load_paired_end_qc_data}}.}

\item{binsize}{A 2-item numeric vector that denotes the bin width for histogram bars. The
first value is for the \code{excluded} plot, and the other value corresponds to the \code{retained}
plot.}

\item{fill}{A 3-item vector of strings that denotes the colors for histogram bars. The first
color is for the \code{excluded} plot, and the other colors correspond to R1 and R2 reads in the
\code{retained} plot.}

\item{alpha}{A numeric. The alpha level for histogram bars.}

\item{ids}{A vector of rownames for subsetting \code{data} for plotting.}

\item{invert}{A logical. Should \code{ids} be used for excluding rows from plotting instead?}
}
\value{
A list with two ggplot objects, \code{excluded} and \code{retained} that correspond to ggplot
objects for reads excluded and retained after Trimmomatic respectively.
}
\description{
Create a histograms for read retention and filtering rates.
}
\details{
Uses output generated by MultiQC on Trimmomatic output for paired-end sequencing data to
produce sample frequency histograms of read retention and filtering rates. Note: this function
only works with Trimmomatic output for paired-end sequencing data.
}
\seealso{
\code{\link{load_paired_end_qc_data}}
}
