% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{pval_histogram}
\alias{pval_histogram}
\title{Plot p-value histogram}
\usage{
pval_histogram(pvalues, bin_width = 0.025, bin_fill = "gray10", alpha = 0.8)
}
\arguments{
\item{pvalues}{A vector of pvalues.}

\item{bin_width}{Size of each histogram bin.}

\item{bin_fill}{Bin fill color.}

\item{alpha}{Bin fill color alpha value.}
}
\value{
A ggplot object.
}
\description{
Plot p-value histogram.
}
