% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{pval_qqplot}
\alias{pval_qqplot}
\title{Quantile-quantile plot}
\usage{
pval_qqplot(
  pvalues,
  outliers = NULL,
  sig_cutoff = 0.05,
  plot_lambda = T,
  df = 1
)
}
\arguments{
\item{pvalues}{A vector of pvalues with associated feature IDs.}

\item{outliers}{A vector of outlier IDs that correspond to the names in 'pvalues'.}

\item{sig_cutoff}{Adjusted p-value significance threshold.}

\item{plot_lambda}{If 'TRUE' calculate the genomic inflation factor and overlay it on the plot.}

\item{df}{Degrees of freedom on the theoretical distribution. Used in calculating the genomic
inflation factor. Only relevant when \code{plot_lambda} is 'TRUE'.}
}
\value{
A ggplot object.
}
\description{
Plots observed vs. expected p-values.
}
\details{
Plot observed p-values vs. expected p-values. Expected p-values are assumed to follow a
uniform distribution.
}
