% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{rint}
\alias{rint}
\title{Rank inverse-normal transform}
\usage{
rint(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
Apply the rank inverse-normal transform (RINT) to a vector.
}
