% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode_utils.R
\name{subset_gencode_gtf}
\alias{subset_gencode_gtf}
\title{Subset GENCODE GTF to specific features}
\usage{
subset_gencode_gtf(gtf, feature_type = c("transcript", "gene", "exon"))
}
\arguments{
\item{gtf}{A data frame with data from an imported GENCODE GTF file.}

\item{feature_type}{A string denoting the feature type for which to retrieve records. This
corresponds to the 3rd column values in a GTF file.}
}
\value{
A data frame in GTF-like format with an additional column (column 10) for the
GENCODE ID.
}
\description{
Creates a data frame with a subset of GTF records based on feature type.
}
\details{
Filters a GTF-format data frame to only records of a specific feature type (based on column 3
of a GTF record). Adds an additional column with the GENCODE ID parsed from column 9.
}
