% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_utils.R
\name{subset_txi}
\alias{subset_txi}
\title{Subsets the tables of a tximport object}
\usage{
subset_txi(txi, ids, dimension)
}
\arguments{
\item{txi}{The tximport object to subset.}

\item{ids}{The list of row or column indices/names to use for subsetting.}

\item{dimensions}{The dimenions to subset. \code{1} for rows and \code{2} for columns.}
}
\value{
A tximport object.
}
\description{
Subsets tximport tables using specified IDs.
}
\details{
Uses a user-specified list of row or column names to subset the \code{counts}, \code{abundance}, and
\code{length} elements of the tximport object. Orders the subset according to the order in the
user-specified list.
}
