% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{mvalue_to_beta}
\alias{mvalue_to_beta}
\title{DNA methylation beta values}
\usage{
mvalue_to_beta(data)
}
\arguments{
\item{data}{A numeric vector or matrix of DNAm array M-values.}
}
\value{
A numeric vector or matrix of beta intensity values.
}
\description{
Convert DNA methylation (DNAm) array M-values to beta intensity values.
}
