% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R, R/dge_utils.R, R/limma.R
\name{per_sample_count_distribution}
\alias{per_sample_count_distribution}
\title{Plot per sample count distribution}
\usage{
per_sample_count_distribution(
  dds,
  normalized = T,
  point_size = 2.5,
  point_alpha = 1,
  y_lim = NULL
)

per_sample_count_distribution(
  dds,
  normalized = T,
  point_size = 2.5,
  point_alpha = 1,
  y_lim = NULL
)

per_sample_count_distribution(
  dds,
  normalized = T,
  point_size = 2.5,
  point_alpha = 1,
  y_lim = NULL
)
}
\arguments{
\item{dds}{A DESeqDataSet object.}

\item{point_size}{Plot point size.}

\item{point_alpha}{Plot point alpha value.}

\item{y_lim}{Plot y-axis range.}
}
\value{
A ggplot object.

A ggplot object.

A ggplot object.
}
\description{
Plot feature count distribution summary statistics per sample.

Plot feature count distribution summary statistics per sample.

Plot feature count distribution summary statistics per sample.
}
\details{
Plots the 25th, 50th, and 75th percentil of the feature count distribution for each sample.

Plots the 25th, 50th, and 75th percentil of the feature count distribution for each sample.

Plots the 25th, 50th, and 75th percentil of the feature count distribution for each sample.
}
