% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R, R/dge_utils.R, R/limma.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(
  dds_results,
  outliers = NULL,
  sig_cutoff = 0.05,
  sig_up_color = "red4",
  sig_down_color = "steelblue3",
  nonsig_color = "gray60",
  outlier_color = "goldenrod3",
  sig_alpha = 0.8,
  nonsig_alpha = 0.4,
  sig_size = 3,
  nonsig_size = 2,
  outlier_size = sig_size * 2.5
)

volcano_plot(
  dds_results,
  outliers = NULL,
  sig_cutoff = 0.05,
  sig_up_color = "red4",
  sig_down_color = "steelblue3",
  nonsig_color = "gray60",
  outlier_color = "goldenrod3",
  sig_alpha = 0.8,
  nonsig_alpha = 0.4,
  sig_size = 3,
  nonsig_size = 2,
  outlier_size = sig_size * 2.5
)

volcano_plot(
  dds_results,
  outliers = NULL,
  sig_cutoff = 0.05,
  sig_up_color = "red4",
  sig_down_color = "steelblue3",
  nonsig_color = "gray60",
  outlier_color = "goldenrod3",
  sig_alpha = 0.8,
  nonsig_alpha = 0.4,
  sig_size = 3,
  nonsig_size = 2,
  outlier_size = sig_size * 2.5
)
}
\arguments{
\item{outliers}{A vector of outlier IDs.}

\item{sig_cutoff}{Adjusted p-value significance threshold.}

\item{sig_up_color}{Point color for statistically significant up-regulated features.}

\item{sig_down_color}{Point color for statistically significant down-regulated features.}

\item{nonsig_color}{Point color for features not statistical significant.}

\item{outlier_color}{Point color for statistically significant features flagged as outliers.}

\item{sig_alpha}{Point color alpha value for statistically significant features.}

\item{nonsig_alpha}{Point color alpha value for features not statistically significant.}

\item{sig_size}{Point size value for statistically significant features.}

\item{nonsig_size}{Point size value for outliers.}
}
\value{
A ggplot object.

A ggplot object.

A ggplot object.
}
\description{
Plot log fold change vs log p-value.

Plot log fold change vs log p-value.

Plot log fold change vs log p-value.
}
\details{
Plots the log2 fold change vs -log10 pvalue for features tested for differential expression.

Plots the log2 fold change vs -log10 pvalue for features tested for differential expression.

Plots the log2 fold change vs -log10 pvalue for features tested for differential expression.
}
